[modify] reformat codes and optimize imports.

+++ Problems/src/SquareRootX_69.java
@@ -4,13 +4,13 @@ import java.io.InputStreamReader;
 
 /**
  * Implement int sqrt(int x).
- *
+ * <p>
  * Compute and return the square root of x, where x is guaranteed to be a
  * non-negative integer.
- *
+ * <p>
  * Since the return type is an integer, the decimal digits are truncated
  * and only the integer part of the result is returned.
- *
+ * <p>
  * 来源：力扣（LeetCode）
  * 链接：https://leetcode-cn.com/problems/sqrtx
  * 著作权归领扣网络所有。商业转载请联系官方授权，非商业转载请注明出处。
@@ -20,7 +20,7 @@ public class SquareRootX_69 {
     public static void main(String[] args) throws IOException {
         BufferedReader in = new BufferedReader(new InputStreamReader(System.in));
         String line;
-        while ((line = in.readLine()) != null) {
+        while((line = in.readLine()) != null) {
             int x = Integer.parseInt(line);
 
             int ret = new SquareRootX_69().mySqrt(x);
@@ -40,10 +40,10 @@ public class SquareRootX_69 {
             return x;
         }
 
-        double xn = (double)x;
+        double xn = x;
         while(xn - x / xn > 1e-5) {
             xn = xn / 2 + x / xn / 2;
         }
-        return (int)xn;
+        return (int) xn;
     }
 }
